+++ palmod/Game/GAME_Bleach_DS.cpp
@@ -8,8 +8,8 @@ stExtraDef* CGame_BLEACH_DS::BLEACH_DS_EXTRA_CUSTOM = nullptr;
 
 CDescTree CGame_BLEACH_DS::MainDescTree = nullptr;
 
-size_t CGame_BLEACH_DS::rgExtraCountAll[BLEACH_DS_NUMUNIT + 1];
-size_t CGame_BLEACH_DS::rgExtraLoc[BLEACH_DS_NUMUNIT + 1];
+uint32_t CGame_BLEACH_DS::rgExtraCountAll[BLEACH_DS_NUMUNIT + 1];
+uint32_t CGame_BLEACH_DS::rgExtraLoc[BLEACH_DS_NUMUNIT + 1];
 
 UINT32 CGame_BLEACH_DS::m_nTotalPaletteCountForBleach = 0;
 UINT32 CGame_BLEACH_DS::m_nExpectedGameROMSize = 0x08000000; // Update to the actual size of the ROM you expect
@@ -64,8 +64,8 @@ CGame_BLEACH_DS::CGame_BLEACH_DS(UINT32 nConfirmedROMSize)
     pButtonLabelSet = DEF_BUTTONLABEL_BLEACH; // Check out the available options in buttondef.h
 
     //Create the redirect buffer
-    rgUnitRedir = new size_t[nUnitAmt + 1];
-    memset(rgUnitRedir, NULL, sizeof(size_t) * nUnitAmt);
+    rgUnitRedir = new uint32_t[nUnitAmt + 1];
+    memset(rgUnitRedir, NULL, sizeof(uint32_t) * nUnitAmt);
 
     //Create the file changed flag
     PrepChangeTrackingArray();
@@ -106,12 +106,12 @@ CDescTree* CGame_BLEACH_DS::GetMainTree()
     return &CGame_BLEACH_DS::MainDescTree;
 }
 
-size_t CGame_BLEACH_DS::GetExtraCt(size_t nUnitId, BOOL bCountVisibleOnly)
+uint32_t CGame_BLEACH_DS::GetExtraCt(uint32_t nUnitId, BOOL bCountVisibleOnly)
 {
     return _GetExtraCount(rgExtraCountAll, BLEACH_DS_NUMUNIT, nUnitId, BLEACH_DS_EXTRA_CUSTOM);
 }
 
-size_t CGame_BLEACH_DS::GetExtraLoc(size_t nUnitId)
+uint32_t CGame_BLEACH_DS::GetExtraLoc(uint32_t nUnitId)
 {
     return _GetExtraLocation(rgExtraLoc, BLEACH_DS_NUMUNIT, nUnitId, BLEACH_DS_EXTRA_CUSTOM);
 }
@@ -144,7 +144,7 @@ sDescTreeNode* CGame_BLEACH_DS::InitDescTree()
     return NewDescTree;
 }
 
-sFileRule CGame_BLEACH_DS::GetRule(size_t nUnitId)
+sFileRule CGame_BLEACH_DS::GetRule(uint32_t nUnitId)
 {
     sFileRule NewFileRule;
 
@@ -157,42 +157,42 @@ sFileRule CGame_BLEACH_DS::GetRule(size_t nUnitId)
     return NewFileRule;
 }
 
-size_t CGame_BLEACH_DS::GetCollectionCountForUnit(size_t nUnitId)
+uint32_t CGame_BLEACH_DS::GetCollectionCountForUnit(uint32_t nUnitId)
 {
     return _GetCollectionCountForUnit(BLEACH_DS_UNITS, rgExtraCountAll, BLEACH_DS_NUMUNIT, BLEACH_DS_EXTRALOC, nUnitId, BLEACH_DS_EXTRA_CUSTOM);
 }
 
-size_t CGame_BLEACH_DS::GetNodeCountForCollection(size_t nUnitId, size_t nCollectionId)
+uint32_t CGame_BLEACH_DS::GetNodeCountForCollection(uint32_t nUnitId, uint32_t nCollectionId)
 {
     return _GetNodeCountForCollection(BLEACH_DS_UNITS, rgExtraCountAll, BLEACH_DS_NUMUNIT, BLEACH_DS_EXTRALOC, nUnitId, nCollectionId, BLEACH_DS_EXTRA_CUSTOM);
 }
 
-LPCWSTR CGame_BLEACH_DS::GetDescriptionForCollection(size_t nUnitId, size_t nCollectionId)
+LPCWSTR CGame_BLEACH_DS::GetDescriptionForCollection(uint32_t nUnitId, uint32_t nCollectionId)
 {
     return _GetDescriptionForCollection(BLEACH_DS_UNITS, BLEACH_DS_EXTRALOC, nUnitId, nCollectionId);
 }
 
-size_t CGame_BLEACH_DS::GetPaletteCountForUnit(size_t nUnitId)
+uint32_t CGame_BLEACH_DS::GetPaletteCountForUnit(uint32_t nUnitId)
 {
     return _GetPaletteCountForUnit(BLEACH_DS_UNITS, rgExtraCountAll, BLEACH_DS_NUMUNIT, BLEACH_DS_EXTRALOC, nUnitId, BLEACH_DS_EXTRA_CUSTOM);
 }
 
-const sGame_PaletteDataset* CGame_BLEACH_DS::GetPaletteSet(size_t nUnitId, size_t nCollectionId)
+const sGame_PaletteDataset* CGame_BLEACH_DS::GetPaletteSet(uint32_t nUnitId, uint32_t nCollectionId)
 {
     return _GetPaletteSet(BLEACH_DS_UNITS, nUnitId, nCollectionId);
 }
 
-const sDescTreeNode* CGame_BLEACH_DS::GetNodeFromPaletteId(size_t nUnitId, size_t nPaletteId, bool fReturnBasicNodesOnly)
+const sDescTreeNode* CGame_BLEACH_DS::GetNodeFromPaletteId(uint32_t nUnitId, uint32_t nPaletteId, bool fReturnBasicNodesOnly)
 {
     return _GetNodeFromPaletteId(BLEACH_DS_UNITS, rgExtraCountAll, BLEACH_DS_NUMUNIT, BLEACH_DS_EXTRALOC, nUnitId, nPaletteId, BLEACH_DS_EXTRA_CUSTOM, fReturnBasicNodesOnly);
 }
 
-const sGame_PaletteDataset* CGame_BLEACH_DS::GetSpecificPalette(size_t nUnitId, size_t nPaletteId)
+const sGame_PaletteDataset* CGame_BLEACH_DS::GetSpecificPalette(uint32_t nUnitId, uint32_t nPaletteId)
 {
     return _GetSpecificPalette(BLEACH_DS_UNITS, rgExtraCountAll, BLEACH_DS_NUMUNIT, BLEACH_DS_EXTRALOC, nUnitId, nPaletteId, BLEACH_DS_EXTRA_CUSTOM);
 }
 
-void CGame_BLEACH_DS::LoadSpecificPaletteData(size_t nUnitId, size_t nPalId)
+void CGame_BLEACH_DS::LoadSpecificPaletteData(uint32_t nUnitId, uint32_t nPalId)
 {
     if (nUnitId != BLEACH_DS_EXTRALOC)
     {
